backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.8.3
google-api-core==2.19.0
python-dotenv==1.0.1
gunicorn==22.0.0